+++ src/components/GameGrid.tsx
@@ -1,31 +1,8 @@
-import { useState, useEffect } from "react";
-import apiClient from "../services/api-client";
 import { Text } from "@chakra-ui/react";
+import useGames from "../hooks/useGames";
 
 const GameGrid = () => {
-  const [games, setGames] = useState<Game[]>([]);
-  const [error, setError] = useState("");
-
-  interface Game {
-    id: number;
-    name: string;
-  }
-
-  interface FetchGamesResponse {
-    count: number;
-    results: Game[];
-  }
-
-  useEffect(() => {
-    apiClient
-      .get<FetchGamesResponse>("/games")
-      .then((res) => {
-        setGames(res.data.results);
-      })
-      .catch((err) => {
-        setError(err.message);
-      });
-  });
+  const { games, error } = useGames();
 
   return (
     <>
